Add tests

+++ src/main/java/by/epam/movierating/command/impl/actor/DeleteActorCommand.java
@@ -2,6 +2,8 @@ package by.epam.movierating.command.impl.actor;
 
 import by.epam.movierating.command.Command;
 import by.epam.movierating.command.constant.ParameterName;
+import by.epam.movierating.command.security.RoleType;
+import by.epam.movierating.command.security.SecurityManager;
 import by.epam.movierating.service.ActorService;
 import by.epam.movierating.service.exception.ServiceException;
 import by.epam.movierating.service.factory.ServiceFactory;
@@ -22,19 +24,20 @@ public class DeleteActorCommand implements Command {
     @Override
     public void execute(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
+        if (SecurityManager.getInstance().checkRoles(request, response, RoleType.ADMIN)) {
+            response.setContentType("text/plain");
 
-        response.setContentType("text/plain");
+            int idActor = Integer.parseInt(request.getParameter(ParameterName.ACTOR_ID));
 
-        int idActor = Integer.parseInt(request.getParameter(ParameterName.ACTOR_ID));
-
-        ServiceFactory serviceFactory = ServiceFactory.getInstance();
-        ActorService actorService = serviceFactory.getActorService();
-        try {
-            boolean result = actorService.deleteActor(idActor);
-            response.getWriter().print(result);
-        } catch (ServiceException e) {
-            logger.error(e);
-            response.getWriter().print(false);
+            ServiceFactory serviceFactory = ServiceFactory.getInstance();
+            ActorService actorService = serviceFactory.getActorService();
+            try {
+                boolean result = actorService.deleteActor(idActor);
+                response.getWriter().print(result);
+            } catch (ServiceException e) {
+                logger.error(e);
+                response.getWriter().print(false);
+            }
         }
     }
 }
