+++ src/main/java/by/epam/movierating/command/impl/admin/DeleteUserCommand.java
@@ -2,6 +2,8 @@ package by.epam.movierating.command.impl.admin;
 
 import by.epam.movierating.command.Command;
 import by.epam.movierating.command.constant.ParameterName;
+import by.epam.movierating.command.security.RoleType;
+import by.epam.movierating.command.security.SecurityManager;
 import by.epam.movierating.service.UserService;
 import by.epam.movierating.service.exception.ServiceException;
 import by.epam.movierating.service.factory.ServiceFactory;
@@ -21,16 +23,18 @@ public class DeleteUserCommand implements Command {
     @Override
     public void execute(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        response.setContentType("text/plain");
-        int isUser = Integer.parseInt(request.getParameter(ParameterName.USER_ID));
-        ServiceFactory serviceFactory = ServiceFactory.getInstance();
-        UserService userService = serviceFactory.getUserService();
-        try {
-            boolean result = userService.deleteUser(isUser);
-            response.getWriter().print(result);
-        } catch (ServiceException e) {
-            logger.error(e);
-            response.getWriter().print(false);
+        if (SecurityManager.getInstance().checkRoles(request, response, RoleType.ADMIN)) {
+            response.setContentType("text/plain");
+            int isUser = Integer.parseInt(request.getParameter(ParameterName.USER_ID));
+            ServiceFactory serviceFactory = ServiceFactory.getInstance();
+            UserService userService = serviceFactory.getUserService();
+            try {
+                boolean result = userService.deleteUser(isUser);
+                response.getWriter().print(result);
+            } catch (ServiceException e) {
+                logger.error(e);
+                response.getWriter().print(false);
+            }
         }
     }
 }
