+++ src/main/java/by/epam/movierating/command/impl/movie/DeleteMovieCommand.java
@@ -2,6 +2,8 @@ package by.epam.movierating.command.impl.movie;
 
 import by.epam.movierating.command.Command;
 import by.epam.movierating.command.constant.ParameterName;
+import by.epam.movierating.command.security.RoleType;
+import by.epam.movierating.command.security.SecurityManager;
 import by.epam.movierating.service.MovieService;
 import by.epam.movierating.service.exception.ServiceException;
 import by.epam.movierating.service.factory.ServiceFactory;
@@ -21,17 +23,19 @@ public class DeleteMovieCommand implements Command {
     @Override
     public void execute(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        response.setContentType("text/plain");
-        int idMovie = Integer.parseInt(request.getParameter(ParameterName.MOVIE_ID));
+        if (SecurityManager.getInstance().checkRoles(request, response, RoleType.ADMIN)) {
+            response.setContentType("text/plain");
+            int idMovie = Integer.parseInt(request.getParameter(ParameterName.MOVIE_ID));
 
-        ServiceFactory serviceFactory = ServiceFactory.getInstance();
-        MovieService movieService = serviceFactory.getMovieService();
-        try {
-            boolean result = movieService.deleteMovie(idMovie);
-            response.getWriter().print(result);
-        } catch (ServiceException e) {
-            logger.error(e);
-            response.getWriter().print(false);
+            ServiceFactory serviceFactory = ServiceFactory.getInstance();
+            MovieService movieService = serviceFactory.getMovieService();
+            try {
+                boolean result = movieService.deleteMovie(idMovie);
+                response.getWriter().print(result);
+            } catch (ServiceException e) {
+                logger.error(e);
+                response.getWriter().print(false);
+            }
         }
     }
 }
